+++ apps/docs/src/app/_components/layout/PageActions/index.ts
@@ -1,0 +1,3 @@
+import PageActions from "@/app/_components/layout/PageActions/PageActions";
+
+export default PageActions;
